perf(SeparatorExpression): Expression 클래스를 만들어서 문자열 입력에 대해 검증 반영
1. StringCalculatorService 애서 코드 이동
2. 이 메소드를 실행하던 Application, Service 수정.
