+++ src/main/java/calculator/CalculatorApplication.java
@@ -1,15 +1,13 @@
 package calculator;
 
+import calculator.domain.CalcNumber;
 import calculator.service.StringCalculatorService;
 import calculator.view.InputView;
-
-import java.util.List;
 
 public class CalculatorApplication {
     public static void main(String[] args) {
         String input = InputView.input();
-        List<String> operations = StringCalculatorService.getOperations(input);
-        int number = StringCalculatorService.calculate(operations).getNumber();
-        System.out.println("number = " + number);
+        CalcNumber result = StringCalculatorService.calculate(input);
+        System.out.println("result = " + result.getNumber());
     }
 }
