+++ src/main/java/calculator/service/StringCalculatorService.java
@@ -2,25 +2,18 @@ package calculator.service;
 
 import calculator.domain.CalcNumber;
 import calculator.domain.CalcOperator;
+import calculator.domain.SeparatorExpression;
 
 import java.util.List;
 
 public class StringCalculatorService {
     private static final int FIRST_INPUT_INDEX = 0;
-    private static final String SEPARATOR = " ";
-    private static final String NO_INPUT_MESSAGE = "입력값이 없습니다.";
 
     private StringCalculatorService() {
     }
 
-    public static List<String> getOperations(String input) {
-        if (isNullOrBlank(input)) {
-            throw new IllegalArgumentException(NO_INPUT_MESSAGE);
-        }
-        return List.of(input.split(SEPARATOR));
-    }
-
-    public static CalcNumber calculate(List<String> operations) {
+    public static CalcNumber calculate(String input) {
+        List<String> operations = SeparatorExpression.getOperations(input);
         CalcNumber result = new CalcNumber(operations.get(FIRST_INPUT_INDEX));
         for (int i = 2; i < operations.size(); i += 2) {
             CalcOperator operator = new CalcOperator(operations.get(i - 1));
@@ -28,9 +21,5 @@ public class StringCalculatorService {
             result = operator.operate().calculate(result, nextNum);
         }
         return result;
-    }
-
-    private static boolean isNullOrBlank(String input) {
-        return input == null || input.isBlank();
     }
 }
